+++ Pandora/Pattern/Functor/Adjoint.hs
@@ -2,11 +2,11 @@ module Pandora.Pattern.Functor.Adjoint where
 
 import Pandora.Core.Functor (type (:.), type (:=))
 import Pandora.Pattern.Category (identity)
-import Pandora.Pattern.Functor.Covariant (Covariant)
+import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 
 type (-|) = Adjoint
 
-infixl 3 -|, |-
+infixl 3 -|, |-, -|$, $|-
 
 {- |
 > When providing a new instance, you should ensure it satisfies the four laws:
@@ -35,3 +35,9 @@ class (Covariant t, Covariant u) => Adjoint t u where
 	-- | Also known as 'counit'
 	epsilon :: t :. u := a -> a
 	epsilon = psi identity
+
+	(-|$) :: Covariant v => v a -> (t a -> b) -> v (u b)
+	x -|$ f = (-| f) <$> x
+
+	($|-) :: Covariant v => v (t a) -> (a -> u b) -> v b
+	x $|- f = (|- f) <$> x